travel-seo-pulse/requirements.txt
feedparser>=6.0
anthropic>=0.40.0
python-substack>=0.2.0
googlenewsdecoder>=0.1.0
